blog html was written and Thymeleaf security add project

+++ src/main/java/uz/blog/dto/response/BlogResponseDto.java
@@ -14,7 +14,7 @@ public class BlogResponseDto extends BaseServerModifierDto {
 
     private String text;
 
-    private Integer numberOfReadings;
+    private long numberOfReadings;
 
     private boolean checked;
